Fix table formatting
The leading spaces were overriding the formatting of the table and pushing that content out into a code block. I've replaced the six leading spaces with an em space and a normal space – this looks visually right on GitHub but you might need to preview it in the actual docs.